:sparkles:  feat: 엿보기 컴포넌트 데이터 외부에서 주입하도록 구현
