+++ src/components/organism/PreviewBottomSheet/index.tsx
@@ -1,4 +1,4 @@
-import { useParams, Link } from 'react-router-dom';
+import { useParams, Link, useLocation } from 'react-router-dom';
 import { Button, Text } from '../../atoms';
 import {
   BottomSheetWrapper,
@@ -9,43 +9,18 @@ import {
   CloseBtn,
 } from './style';
 
-const dummyPost = {
-  postId: 0,
-  title: '제목입니당',
-  previewText: '엿보기 문구입니당',
-  availableAt: 'yyyy-MM-dd',
-  state: 'Closed',
-  location: [
-    {
-      latitude: '37.491837217869616',
-      longitude: '127.02959879978368',
-    },
-  ],
-  authorityCount: 3,
-  authorityImageList: [
-    {
-      imageOrder: 0,
-      authorityId: 0,
-      imageUrl: '',
-    },
-    {
-      imageOrder: 0,
-      authorityId: 0,
-      imageUrl: '',
-    },
-  ],
-};
 const PreviewBottomSheet = () => {
   const { id } = useParams();
+  const { state } = useLocation();
   return (
     <BottomSheetWrapper>
-      <FilmTitle textType="Heading3">{dummyPost.title}</FilmTitle>
+      <FilmTitle textType="Heading3">{state.title}</FilmTitle>
       <FilmPreviewText textType="Paragraph1">
-        #{id}필름의 {dummyPost.previewText}
+        #{id}필름의 {state.previewText}
       </FilmPreviewText>
       <DateWrapper>
         <Text textType="Heading4">사진 나오는 날</Text>
-        <FilmOpenDate textType="Paragraph2">{dummyPost.availableAt}</FilmOpenDate>
+        <FilmOpenDate textType="Paragraph2">{state.availableAt}</FilmOpenDate>
       </DateWrapper>
       <Button buttonType="PrimaryBtn" width="100%">
         필름 삭제하기
